Use quasar.api.table.TableColumn instead of custom type in destination API
